+++ src/CosmosDB/CosmosDB/Helpers/Constants.cs
@@ -160,6 +160,12 @@ namespace Microsoft.Azure.Commands.CosmosDB.Helpers
         public const string ResetPartitionThroughputLayoutHelpMessage = "Set this switch to reset the throughput layout for all partitions.";
         public const string PhysicalPartitionThroughputInfosHelpMessage = "List of PhysicalPartitionInfo objects for which throughput needs to be retrieved..";
         public const string PhysicalPartitionIdsHelpMessage = "List of PhysicalPartitionId for which throughput needs to be retrieved.";
+        public const string VectorEmbeddingPolicyVectorEmbeddingHelpMessage = "Represents a vector embedding. A vector embedding is used to define a vector field in the documents.";
+        public const string VectorEmbeddingPathHelpMessage = "The path to the vector field in the document.";
+        public const string VectorEmbeddingDataTypeHelpMessage = "Indicates the data type of vector.";
+        public const string VectorEmbeddingDistanceFunctionHelpMessage = "The distance function to use for distance calculation in between vectors.";
+        public const string VectorEmbeddingDimensionsHelpMessage = "The number of dimensions in the vector.";
+        public const string IndexingPolicyVectorIndexHelpMessage = "Array of strings containing path to the vector field in the document.";
 
         //SQL Client Side Encryption
         public const string ClientEncryptionKeyName = "Client Encryption Key name.";
@@ -231,6 +237,11 @@ namespace Microsoft.Azure.Commands.CosmosDB.Helpers
         public const string ThroughputHelpMessage = "Throughput value in int.";
         public const string AutoscaleMaxThroughputHelpMessage = "Maximum Throughput value if autoscale is enabled.";
         public const string ThroughputTypeHelpMessage = "Throughput type to migrate to. Possible values are: Autoscale, Manual.";
+
+        // Throughput bucketing help messages
+        public const string ThroughputBucketsObjectHelpMessage = "Array of throughput buckets to be applied to the Cosmos DB resource.";
+        public const string ThroughputBucketIdHelpMessage = "The ID of the throughput bucket.";
+        public const string ThroughputBucketMaxThroughputHelpMessage = "The maximum throughput of the throughput bucket.";
 
         // Role cmdlets help messages
         public const string PrincipalIdHelpMessage = "Object ID (Guid) of the AAD principal to which the Role Assignment is being granted. This could be user, group, service principal, or managed identity.";